C# lower aligns with python structure
